bionic: timezone update to tzdata2012b and fix 32 bit alignment
The update to Android 4.0.4 updated the timezone data file to once again
not be 32 bit aligned, causing SIGBUS on ARMv5/6.

The zoneinfo.dat file is memory mapped and then read as an int array.
On some platforms this is causing alignment errors (SIGBUS) because
the records are not 32 bit aligned.

Change-Id: I1ed0e7842e6615bf0123f258a21bc40e838fd18e